+++ README.md
@@ -29,7 +29,11 @@ Have a look at the `conf/application-conf` file and fill in sections for
  - **LDAP** (if using; defaulting to in-memory), 
  - **SMTP** (if using; defaulting to a mocked smtp that does not send any emails.)
 
-Run the app with `sbt run`, run tests with `sbt test` and integration tests with `sbt it:test`
+Then proceed with:
+
+* Run the app with `sbt run`
+* Run tests with `sbt test`
+* Run LDAP-integration tests with `sbt ldap:test`
 
 ## Functionality status
