--- it/com/yetu/oauth2provider/services/LDAPBaseSpec.scala
@@ -1,9 +1,0 @@
-package com.yetu.oauth2provider.services
-
-import com.yetu.oauth2provider.registry.IntegrationTestRegistry
-import org.scalatest.BeforeAndAfterEach
-import org.scalatestplus.play.{OneAppPerSuite, PlaySpec}
-
-class LDAPBaseSpec extends PlaySpec with OneAppPerSuite with IntegrationTestRegistry with DefaultTestVariables with BeforeAndAfterEach {
-
-}
