+++ test/com/yetu/oauth2provider/ldap/LDAPPermissionsService.scala
@@ -1,0 +1,18 @@
+package com.yetu.oauth2provider.ldap
+
+//TODO: implement permissions correctly and make sure this test leaves no traces behind.
+class LDAPPermissionsService extends LDAPBaseSpecITSpec {
+  //  "The LDAP permission service" must {
+  //    val clientPermission = ClientPermission("123456", Some(List("scope1")))
+  //    "delete, store and retrieve a permissions " in {
+  //      personService.deleteUser(testUser.userId)
+  //      permissionService.deletePermission(testUser.userId, clientPermission.clientId)
+  //      personService.save(testUser.toBasicProfile, SaveMode.SignUp)
+  //      permissionService.savePermission(testUser.userId, clientPermission, true)
+  //      val retrieved = permissionService.findPermission(testUser.userId, clientPermission.clientId)
+  //      retrieved.get mustEqual clientPermission
+  //      permissionService.deletePermission(testUser.userId, clientPermission.clientId)
+  //      personService.deleteUser(testUser.userId)
+  //    }
+  //  }
+}
