+++ test/com/yetu/oauth2provider/registry/IntegrationTestRegistry.scala
@@ -1,6 +1,5 @@
 package com.yetu.oauth2provider.registry
 
-
 /**
  * use the real LDAP for these integration tests.
  */
